refactor(user-management): move byte-identical team_service to autobot_shared (#12647)
team_service.py was byte-identical across autobot-backend and
autobot-slm-backend — 701 duplicated lines, so a fix landed in one service
or the other but never both by accident.

It is the first user_management *service* that could move: its entire
runtime dependency set already lives in autobot_shared — Team/TeamMembership
and the audit vocabulary (#13130), BaseService/TenantContext (#12972). Its
only mentions of User and Organization are in docstrings and error strings,
never as classes, so it does not need the concrete per-backend models.

Placed flat as autobot_shared/user_management/team_service.py, matching the
existing base_service.py rather than inventing a second services/ location.

Both backends keep a re-export shim so all existing importers, including
each services/__init__.py, are untouched.

organization_service.py is also byte-identical but deliberately not moved:
it imports the concrete User and Organization at runtime, and those stay
backend-local by design.
